requirements.txt
databricks-sdk>=0.40.0
pydantic>=2.0.0
